move pandas_time_delta_to_quantity to time module
