+++ weldx/tests/test_time.py
@@ -1,5 +1,6 @@
 """Test the `Time` class."""
 
+import math
 from typing import List, Tuple, Type, Union
 
 import numpy as np
@@ -9,7 +10,7 @@ import xarray as xr
 from pandas import DatetimeIndex, Timedelta, TimedeltaIndex, Timestamp
 
 from weldx import Q_
-from weldx.time import Time
+from weldx.time import Time, pandas_time_delta_to_quantity
 
 
 def _initialize_delta_type(cls_type, values, unit):
@@ -322,3 +323,36 @@ class TestTime:
 
         arr2 = time.as_data_array().weldx.time_ref_restore()
         assert arr.time.identical(arr2.time)
+
+
+# test_pandas_time_delta_to_quantity ---------------------------------------------------
+
+
+def test_pandas_time_delta_to_quantity():
+    """Test the 'pandas_time_delta_to_quantity' utility function."""
+    is_close = np.vectorize(math.isclose)
+
+    def _check_close(t1, t2):
+        assert np.all(is_close(t1.magnitude, t2.magnitude))
+        assert t1.units == t2.units
+
+    time_single = pd.TimedeltaIndex([1], unit="s")
+
+    _check_close(pandas_time_delta_to_quantity(time_single), Q_(1, "s"))
+    _check_close(pandas_time_delta_to_quantity(time_single, "ms"), Q_(1000, "ms"))
+    _check_close(pandas_time_delta_to_quantity(time_single, "us"), Q_(1000000, "us"))
+    _check_close(pandas_time_delta_to_quantity(time_single, "ns"), Q_(1000000000, "ns"))
+
+    time_multi = pd.TimedeltaIndex([1, 2, 3], unit="s")
+    _check_close(pandas_time_delta_to_quantity(time_multi), Q_([1, 2, 3], "s"))
+    _check_close(
+        pandas_time_delta_to_quantity(time_multi, "ms"), Q_([1000, 2000, 3000], "ms")
+    )
+    _check_close(
+        pandas_time_delta_to_quantity(time_multi, "us"),
+        Q_([1000000, 2000000, 3000000], "us"),
+    )
+    _check_close(
+        pandas_time_delta_to_quantity(time_multi, "ns"),
+        Q_([1000000000, 2000000000, 3000000000], "ns"),
+    )
