+++ weldx/util.py
@@ -22,7 +22,6 @@ from pint import DimensionalityError
 from scipy.spatial.transform import Rotation as Rot
 from scipy.spatial.transform import Slerp
 
-from .constants import WELDX_QUANTITY as Q_
 from .constants import WELDX_UNIT_REGISTRY as ureg
 
 
@@ -377,32 +376,6 @@ def to_pandas_time_index(
     raise TypeError(
         f"Could not convert {_input_type} " f"to pd.DatetimeIndex or pd.TimedeltaIndex"
     )
-
-
-def pandas_time_delta_to_quantity(
-    time: pd.TimedeltaIndex, unit: str = "s"
-) -> pint.Quantity:
-    """Convert a `pandas.TimedeltaIndex` into a corresponding `pint.Quantity`.
-
-    Parameters
-    ----------
-    time : pandas.TimedeltaIndex
-        Instance of `pandas.TimedeltaIndex`
-    unit :
-        String that specifies the desired time unit.
-
-    Returns
-    -------
-    pint.Quantity :
-        Converted time quantity
-
-    """
-    # from pandas Timedelta documentation: "The .value attribute is always in ns."
-    # https://pandas.pydata.org/pandas-docs/version/0.23.4/generated/pandas.Timedelta.html
-    nanoseconds = time.values.astype(np.int64)
-    if len(nanoseconds) == 1:
-        nanoseconds = nanoseconds[0]
-    return Q_(nanoseconds, "ns").to(unit)
 
 
 def matrix_is_close(mat_a, mat_b, abs_tol=1e-9) -> bool:
